Flatten refactor workspace: eliminate redundant src/ nesting
This commit removes the unnecessary nested src/ directory from the
refactor workspace, aligning it with the target monorepo architecture.

Changes:
- Move src/refactor/src/engine/ → src/refactor/engine/
- Update src/refactor/index.ts to import directly from ./engine/
- Update test imports from ../src/index.js to ../index.js
- Update tsconfig.json include pattern to match new structure
- Remove empty src/refactor/src/ directory

Benefits:
- Reduces directory nesting from 4 to 3 levels
- Simplifies import paths
- Establishes pattern for future workspace flattening
- Maintains all public APIs and test coverage (135 tests pass)

This is the first of 8 workspaces to be flattened. Subsequent PRs
will address the remaining workspaces following this pattern.
