+++ src/refactor/index.ts
@@ -1,1 +1,20 @@
-export * as Refactor from "./src/index.js";
+export {
+    RefactorEngine,
+    WorkspaceEdit,
+    createRefactorEngine
+} from "./engine/index.js";
+export type {
+    ParserBridge,
+    SemanticAnalyzer,
+    WorkspaceReadFile,
+    WorkspaceWriteFile,
+    HotReloadUpdate,
+    ExecuteRenameRequest,
+    ExecuteBatchRenameRequest,
+    RenameRequest,
+    TranspilerPatch,
+    RenameImpactAnalysis,
+    ValidationSummary,
+    ValidateRenameRequestOptions,
+    HotReloadSafetySummary
+} from "./engine/index.js";
