--- src/refactor/src/index.ts
@@ -1,20 +1,0 @@
-export {
-    RefactorEngine,
-    WorkspaceEdit,
-    createRefactorEngine
-} from "./engine/index.js";
-export type {
-    ParserBridge,
-    SemanticAnalyzer,
-    WorkspaceReadFile,
-    WorkspaceWriteFile,
-    HotReloadUpdate,
-    ExecuteRenameRequest,
-    ExecuteBatchRenameRequest,
-    RenameRequest,
-    TranspilerPatch,
-    RenameImpactAnalysis,
-    ValidationSummary,
-    ValidateRenameRequestOptions,
-    HotReloadSafetySummary
-} from "./engine/index.js";
